NMRTools.jl

NMRTools.jl is a simple library for importing and plotting NMR data. Documentation is incomplete and interested users are recommended to check out the Examples.

Note

This package is under active development and it may be a while before its features and syntax stabilises.

NMRTools.ExponentialWindow

NMRTools.GaussWindow

NMRTools.HasPseudoDimension

NMRTools.NMRData

NMRTools.NMRData

NMRTools.NullWindow

NMRTools.SineWindow

NMRTools.UnknownWindow

NMRTools.closest

NMRTools.estimatenoise!

NMRTools.estimatenoise!

NMRTools.fitdiffusion

NMRTools.fitexp

NMRTools.haspseudodimension

NMRTools.loadnmr

NMRTools.loadnmrpipe1d

NMRTools.loadnmrpipe2d

NMRTools.loadnmrpipe3d

NMRTools.metadatahelp

NMRTools.parsenmrpipeheader

NMRTools.scale

NMRTools.ExponentialWindow — Type

ExponentialWindow(lb, tmax)

EM: Exponential Multiply Window. -lb expHz [0.0] Exponential Broaden, Hz. (Q1)

NMRTools.GaussWindow — Type

GaussWindow(expHz, gaussHz, center, tmax)

GM: Lorentz-to-Gauss Window. -g1 expHz [0.0] Inverse Exponential, Hz. (Q1) -g2 gausHz [0.0] Gaussian Width, Hz. (Q2) -g3 center [0.0] Center, 0 to 1. (Q3)

NMRTools.HasPseudoDimension — Type

@traitdef HasPseudoDimension{D}

A trait indicating whether the data object has a non-frequency domain dimension.

Example

@traitfn f(x::X) where {X; HasPseudoDimension{X}} = "Fake!"
@traitfn f(x::X) where {X; Not{HasPseudoDimension{X}}} = "The real deal!"

NMRTools.NMRData — Type

NMRData(data, dims, refdims, name)

The main structure for holding MNR data. A subtype of AbstractDimensionalArray, which maintains and updates its dimensions through transformations and moves dimensions to refdims after reducing operations (like e.g. mean).

NMRTools.NMRData — Type

NMRData(data, dims::Tuple [, name::String]; refdims=(), metadata=nothing)

Constructor with optional name, and keyword refdims and metadata.

Example:

using Dates, DimensionalData

ti = (Ti(DateTime(2001):Month(1):DateTime(2001,12)),
x = X(10:10:100))
A = NMRData(rand(12,10), (ti, x), "example")

julia> A[X(Near([12, 35])), Ti(At(DateTime(2001,5)))];

julia> A[Near(DateTime(2001, 5, 4)), Between(20, 50)];

NMRTools.NullWindow — Type

NullWindow(tmax)

No apodization applied.

NMRTools.SineWindow — Type

SineWindow(offset, endpoint, power, tmax)

SP: Adjustable Sine Window. [SINE] -off offset [0.0] Sine StartPI. (Q1) -end end [1.0] Sine EndPI. (Q2) -pow exp [1.0] Sine Exponent. (Q3)

NMRTools.UnknownWindow — Type

UnknownWindow(tmax)

Unknown apodization applied.

NMRTools.estimatenoise! — Method

estimatenoise(spectra::Array{<:NMRData,1})

Estimate the rms noise level and update :noise metadata for a list of spectra.

NMRTools.estimatenoise! — Method

estimatenoise(d::NMRData)

Estimate the rms noise level in the data and update :noise metadata.

NMRTools.fitdiffusion — Method

fitdiffusion(A::NMRData{T,2}, selector; δ=0.004, Δ=0.1, σ=0.9, Gmax=0.55)

Examples

y = loadnmr("exampledata/pseudo2D_XSTE/test.ft1")
y = settval(y, LinRange(0.05, 0.95, 10))
rH, D = fitdiffusion(y, Between(7.5,9))

NMRTools.fitexp — Method

fitexp(A::NMRData{T,2}, selector)

Examples

y = loadnmr("exampledata/pseudo2D_XSTE/test.ft1")
y = settval(y, LinRange(0.05, 0.95, 10))
rate, plt = fitexp(y, Between(7.5,9))
display(plt)

NMRTools.haspseudodimension — Method

haspseudodimension(spectrum)

Return true if the spectrum contains a non-frequency domain dimension.

Example

julia> y2=loadnmr("exampledata/2D_HN/test.ft2");
julia> haspseudodimension(y2)
false
julia> y3=loadnmr("exampledata/pseudo3D_HN_R2/ft/test%03d.ft2");
julia> haspseudodimension(y3)
true

NMRTools.loadnmr — Function

loadnmr(template::String)

Main function for loading nmrPipe or bruker format NMR data.

Returns an NMRData array, or throws an NMRToolsException is there is a problem.

Examples

nmrPipe import:

loadnmr("exampledata/1D_1H/test.ft1");
loadnmr("exampledata/1D_19F/test.ft1");
loadnmr("exampledata/2D_HN/test.ft2");
loadnmr("exampledata/pseudo2D_XSTE/test.ft1");
loadnmr("exampledata/pseudo3D_HN_R2/ft/test%03d.ft2");

bruker import:

loadnmr("exampledata/1D_19F/pdata/1/1r");

NMRTools.metadatahelp — Method

metadatahelp(entry::Symbol)

Get a brief description of what's represented by metadata entry.

Examples

metadatahelp(:FDQUADFLAG);

NMRTools.scale — Method

scale(d::NMRData)

Return a scaling factor for the data combining the number of scans, receiver gain, and, if specified, the sample concentration.

NMRTools.closest — Method

closest(A, x)

Return the index of the array entry closest in value to x. If there are multiple closest entries, then the first one will be returned.

Example

julia> closest([-1, -0.5, 0.5], 0)
2

julia> closest([1 2 3
                4 5 6], 5.8)
CartesianIndex(2, 3)

NMRTools.loadnmrpipe1d — Method

loadnmrpipe1d(filename, md, mdax)

Return an NMRData array containing spectrum and associated metadata.

NMRTools.loadnmrpipe2d — Method

loadnmrpipe2d(filename, md, mdax)

Return NMRData containing spectrum and associated metadata.

NMRTools.loadnmrpipe3d — Method

loadnmrpipe3d(filename, md, mdax)

Return NMRData containing spectrum and associated metadata.

NMRTools.parsenmrpipeheader — Method

parsenmrpipeheader(header)

Pass a 512 x 4 byte array containing an nmrPipe header file, and returns dictionaries of metadata.

Return values

md: dictionary of spectrum metadata

mdax: array of dictionaries containing axis metadata

Examples

md, mdax = parsenmrpipeheader(header)