Coherences and isotope data

NMRTools defines commonly used nuclei and provides a framework for identifying single- and multiple-quantum coherences associated with them. Reference data on their gyromagnetic ratios, and spin quantum numbers, is also defined and accessible through simple functions.

Nuclei

NMRTools.NMRBase.Nucleus — Type

Nucleus

Enumeration of common nuclei associated with biomolecular NMR. Nuclei are named e.g. H1, C13.

Defined nuclei: H1, H2, C12, C13, N14, N15, F19, P31.

See also spin, gyromagneticratio, Coherence.

Coherences

NMRTools.NMRBase.Coherence — Type

Coherence

Abstract supertype for representing coherences.

See also SQ, MQ.

NMRTools.NMRBase.SQ — Type

SQ(nucleus::Nucleus, label=="")

Representation of a single quantum coherence on a given nucleus.

See also Nucleus, MQ.

NMRTools.NMRBase.MQ — Type

MQ(coherences, label=="")

Representation of a multiple-quantum coherence. Coherences are specified as a tuple of tuples, of the form (nucleus, coherenceorder)

Examples

julia> MQ(((H1,1), (C13,-1)), "ZQ")
MQ(((H1, 1), (C13, -1)), "ZQ")

julia> MQ(((H1,3), (C13,1)), "QQ")
MQ(((H1, 3), (C13, 1)), "QQ")

See also Nucleus, SQ.

NMRTools.NMRBase.coherenceorder — Function

coherenceorder(coherence)

Calculate the total coherence order.

Examples

julia> coherenceorder(SQ(H1))
1

julia> coherenceorder(MQ(((H1,1),(C13,1))))
2

julia> coherenceorder(MQ(((H1,1),(C13,-1))))
0

julia> coherenceorder(MQ(((H1,3),(C13,1))))
4

julia> coherenceorder(MQ(((H1,0),)))
0

See also Nucleus, SQ, MQ.

Reference data

NMRTools.NMRBase.spin — Function

spin(n::Nucleus)

Return the spin quantum number of nucleus n, or nothing if not defined.

Examples

julia> spin(H1)
1//2

See also Coherence.

NMRTools.NMRBase.gyromagneticratio — Function

gyromagneticratio(n::Nucleus)
gyromagneticratio(c::Coherence)

Return the gyromagnetic ratio in Hz/T of a nucleus, or calculate the effective gyromagnetic ratio of a coherence. This is equal to the product of the individual gyromagnetic ratios with their coherence orders.

Returns nothing if not defined.

Examples

julia> gyromagneticratio(H1)
2.6752218744e8

julia> gyromagneticratio(SQ(H1))
2.6752218744e8

julia> gyromagneticratio(MQ(((H1,1),(C13,1))))
3.3480498744e8

julia> gyromagneticratio(MQ(((H1,0),)))
0.0

See also Nucleus, Coherence.